Add senior design project to "Experience" section
I'm a bit torn on where this should go.  It fits in the most with
"Experience" in both layout (simply a description and a bulleted list of
responsibilities), and I like that it fits in with the timeline.  I
don't however like that it makes it hard to fit things on the front
page.

Senior Design will stay in "Experience" for now, but may move in the
future.